feat(ApplicationPage): Extract HeaderActions from Header
This component is only responsible for applications and konnectors. It removes certain functionalities for additional apps, which will be added again in the next commit
